feat(bots): 重构 MoneyBot 并添加新功能
-重构 MoneyBot 类,优化命令处理逻辑
- 新增用户信息保存功能
- 改进快速记账功能,支持规则匹配和金额提取
- 优化数据库表结构,改为 IF NOT EXISTS 方式创建
- 更新 Account、Bill、Category、Rule、Tag 和 User 实体类
- 调整相关 Convert 接口

+++ src/main/java/com/tony/log4m/pojo/dto/RuleDTO.java
@@ -15,19 +15,19 @@ import java.math.BigDecimal;
 @EqualsAndHashCode(callSuper = false)
 public class RuleDTO {
 
-    private Integer id;
+    private Long id;
 
     private String name;
 
-    private Integer userId;
+    private Long userId;
 
-    private Integer accountId;
+    private Long accountId;
 
     private String transactionType;
 
-    private Integer categoryId;
+    private Long categoryId;
 
-    private Integer tagId;
+    private Long tagId;
 
     private BigDecimal amount;
